Add files via upload
This is a path finder app that uses different agents to draw the best possible route between two points. The ghost agent just draws a straight line, ignoring obstacles. The A* agent uses an A* search agent and the Greedy best agent uses a greedy best first search. All the nodes, or points, that were evaluated and not used are highlighted in yellow.