Add files via upload
This is a spin on a connect four game, where the rules are the same but the size of the board can be altered. The user plays against a minimax agent with alpha beta pruning. After each move the console shows the total nodes of the search tree searched and the time it took in seconds